+++ components/sections/organizations/Organization/Organization.scss
@@ -1,29 +1,42 @@
-@import "../../../../styles/fonts.scss";
+@import "styles/fonts.scss";
+@import "styles/sizes.scss";
 
 .organization {
   display: flex;
   padding-left: 5vw;
   padding-right: 5vw;
   justify-content: space-between;
+  flex-wrap: wrap;
+  flex-direction: row;
+
+  @media (max-width: $mobile-max) {
+    flex-direction: column;
+    flex-wrap: nowrap;
+  }
 }
 
-.organizationName {
+.team-member {
+  width: 50%;
+
+  @media (max-width: $mobile-max) {
+    width: 100%;
+  }
+}
+
+.organization-name {
   font-family: $poppins-regular;
   padding-bottom: 20px;
-  font-size: 18px;
-  @media (max-width: 500px) {
-    font-size: 10px;
+  @media (min-width: $desktop-min + 1px) {
+    font-size: 18px;
   }
+  font-size: 15px;
   padding-left: 5vw;
   padding-right: 5vw;
 }
 
-.leftMembers {
-  flex: 1 1 0px;
-}
-
-.rightMembers {
+.members {
   flex: 1 1 0px;
   display: flex;
+  flex-direction: column;
   justify-content: flex-end;
 }
