+++ components/sections/organizations/Organization/index.ts
@@ -1,2 +1,3 @@
-import { OrganizationContainer } from "./Organization";
+import { OrganizationMobileContainer } from "./OrganizationMobile/OrganizationMobile";
+import { OrganizationContainer } from "./Organization.container";
 export default OrganizationContainer;
