+++ components/sections/organizations/Organizations.scss
@@ -1,5 +1,5 @@
-@import "../../../styles/fonts.scss";
-@import "../../../styles/colors.scss";
+@import "styles/fonts.scss";
+@import "styles/colors.scss";
 
 .line {
   border-top: 2px solid $gold-drop;
@@ -9,7 +9,6 @@
 .organizationWrapper {
   margin-top: 20px;
 }
-
 
 .section {
   display: flex;
