+++ components/sections/organizations/TeamMember/TeamMember.scss
@@ -1,4 +1,5 @@
-@import "../../../../styles/fonts.scss";
+@import "styles/fonts.scss";
+@import "styles/sizes.scss";
 
 .member {
   font-family: $poppins-regular;
@@ -8,6 +9,7 @@
   color: #3e693d;
   font-size: 25px;
   margin-bottom: 2px;
+  font-family: $poppins-semibold;
 }
 
 .position {
@@ -15,15 +17,16 @@
   margin-bottom: 10px;
 }
 
-@media (max-width: 500px) {
+@media (max-width: $mobile-max) {
   .name {
     color: #3e693d;
-    font-size: 15px;
+    font-size: 20px;
     margin-bottom: 2px;
+    font-family: $poppins-semibold;
   }
 
   .position {
-    font-size: 10px;
+    font-size: 14px;
     margin-bottom: 10px;
   }
 }
